Permission error on macOS, small improvements
On macOS, a special permission is needed to access the mail directory.
This commit adds support to help users to add this permission with
a screenshot and guides.

This was a bit involved because the image should only be loaded on macOS
and because loading images in egui is quite a pain.